requirements.txt
kafka-python
faker
requests
dlt
psycopg2-binary
google-cloud-bigquery
duckdb
pyarrow